Close #49 - Removed files will slowly fade out
When a file has been deleted from the git repository it will no longer
be removed from the graph instantly. Instead it will be marked for
deletion and slowly fade out.

When it is fully invisible it will be removed from the graph (from the
FileManager and its parent node).

This will only happen in "normal" playback mode. When the user uses a
fast forward the files still will be deleted immediately.

+++ src/graph/Graph.lua
@@ -159,7 +159,11 @@ function Graph.new(ewidth, showLabels)
         if modifier == MOD_ADD then
             modifiedFile = targetNode:addFile(filename);
         elseif modifier == MOD_DELETE then
-            modifiedFile = targetNode:removeFile(filename);
+            if mode == 'normal' then
+                modifiedFile = targetNode:markFileForDeletion(filename);
+            else
+                modifiedFile = targetNode:removeFile(filename);
+            end
         elseif modifier == MOD_MODIFY then
             modifiedFile = targetNode:modifyFile(filename);
         end
